+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.3.0
+
+### [0.3.0](https://github.com/auth0-lab/fga-go-sdk/compare/v0.2.2...v0.3.0) (2022-03-17)
+
+#### Changes
+- chore!: change method names used to set request body (BREAKING CHANGE)
+- chore: fix user agent header format
+- feat: add support for the Watch API
+
 ## v0.2.2
 
 ### [0.2.2](https://github.com/auth0-lab/fga-gp-sdk/compare/v0.2.1...v0.2.2) (2022-03-09)